+++ packages/geo-diary/src/components/ui/map-panel/map-panel.tsx
@@ -1,6 +1,6 @@
 import React, { Dispatch, SetStateAction } from 'react'
 import { mapPanelContainer } from './__styles__'
-import { RouteInformation } from '../map/map'
+import { RouteInformation } from '../map/google-map-component'
 import { Button, isMobile, isIOS, ButtonGroup } from '@reapit/elements'
 import { AppState, useAppState } from '../../../core/app-state'
