+++ packages/geo-diary/src/components/ui/map/types.ts
@@ -1,0 +1,8 @@
+import { Dispatch, SetStateAction } from 'react'
+import { AppState, MapRefs } from '../../../core/app-state'
+
+export interface HandleSetMapParams {
+  mapRefs: MapRefs
+  appState: AppState
+  setAppState: Dispatch<SetStateAction<AppState>>
+}
